+++ tmge/src/main/java/edu/uci/tmge/Tile.java
@@ -2,29 +2,29 @@ package edu.uci.tmge;
 
 public class Tile {
 
-  private double x;
-  private double y;
+  private int x;
+  private int y;
   private int type;
 
-  public Tile(double x, double y, int type) {
+  public Tile(int x, int y, int type) {
     this.x = x;
     this.y = y;
     this.type = type;
   }
 
-  public double getX() {
+  public int getX() {
     return x;
   }
 
-  public void setX(double x) {
+  public void setX(int x) {
     this.x = x;
   }
 
-  public double getY() {
+  public int getY() {
     return y;
   }
 
-  public void setY(double y) {
+  public void setY(int y) {
     this.y = y;
   }
